시설관리 패키지 추가
시설관리 패키지 추가

+++ artpartM/src/main/java/org/ict/artpartM/emp/model/dto/EmpDto.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 public class EmpDto {
-    private int empIdx;
+    private long empIdx;
     private String empId;
     private String empPermanentId;
     private String empJobId;
